requirements.txt
pyswitchbot
blinker
prometheus-client
PyYAML
requests
pytest
pytest-mock
pytest-asyncio
pytest-cov
pytimeparse2